Update binary log read/write structure.
In general the idea is to have a single `tic()` in a main `.m` script. This is passed to classes like `typewriter.Prompter` or `cursor.Cursor` so that they can then log accurate binary timestamps to respective `.typewriterdata` or `.cursordata` binaries. Because the alignment tic is from a single source, post-hoc alignment should therefore be straightforward.